requirements.txt
google-adk==1.0.0
google-generativeai==0.3.2
pydantic==2.6.1
rich
streamlit 